LayoutTests/fast/events/pageshow-pagehide-on-back-cached-with-frames-expected.txt

CONSOLE MESSAGE: ***Top level frame being parsed for the initial page load***
CONSOLE MESSAGE: Subsubframe window.onload
CONSOLE MESSAGE: Subsubframe window.onpageshow, target = [object HTMLDocument], persisted = false
CONSOLE MESSAGE: Subframe window.onload
CONSOLE MESSAGE: Subframe window.onpageshow, target = [object HTMLDocument], persisted = false
CONSOLE MESSAGE: Subsubframe2 window.onload
CONSOLE MESSAGE: Subsubframe2 window.onpageshow, target = [object HTMLDocument], persisted = false
CONSOLE MESSAGE: Subframe2 window.onload
CONSOLE MESSAGE: Subframe2 window.onpageshow, target = [object HTMLDocument], persisted = false
CONSOLE MESSAGE: Main frame window.onload
CONSOLE MESSAGE: Main frame window.onpageshow, target = [object HTMLDocument], persisted = false
CONSOLE MESSAGE: ***Navigating top-level frame to a page that will immediately navigate back to this one***
CONSOLE MESSAGE: Main frame window.onpagehide, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: Subframe window.onpagehide, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: Subsubframe window.onpagehide, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: Subframe2 window.onpagehide, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: Subsubframe2 window.onpagehide, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: Subsubframe window.onpageshow, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: Subframe window.onpageshow, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: Subsubframe2 window.onpageshow, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: Subframe2 window.onpageshow, target = [object HTMLDocument], persisted = true
CONSOLE MESSAGE: Main frame window.onpageshow, target = [object HTMLDocument], persisted = true
Test pageshow/pagehide event behavior when navigating away from a page with frames, putting the page in the page cache, then back to it.
